Embed Parted class into TravelAdviceEdition
